fix(build): resolve strict C89 compiler errors and C4306 warnings
- Fixed unterminated string literals across http_curl, http_aria2,
  http_libuv, http_msh3, and http_raw that broke strict GCC/Clang
  builds due to improperly escaped outer double quotes.
- Stripped #include <windows.h> from http_wininet.c and http_winhttp.c
  to adhere to strict binary bloat prevention mandates.
- Addressed C4306 MSVC 2005 warnings in test headers by properly casting
  integer mock values to pointers via size_t/uintptr_t intermediate casts,
  or substituting with NULL.
- Ran clang-format across all files to normalize style.

+++ tests/test_coroutine.h
@@ -316,13 +316,14 @@ TEST test_coroutine_edge_cases(void) {     /* LCOV_EXCL_LINE */
    */
 #if !defined(_WIN32) && !defined(__APPLE__) && !defined(__EMSCRIPTEN__)
   ASSERT_EQ(C_ABSTRACT_HTTP_ERR_NOMEM,
-            abstract_http_coroutine_init(&co, 0, (abstract_http_coroutine_cb)1,
-                                         NULL));
+            abstract_http_coroutine_init(
+                &co, 0, (abstract_http_coroutine_cb)(size_t)1, NULL));
 #else
   {
     int edge_rc =
         abstract_http_coroutine_init(/* LCOV_EXCL_LINE */
-                                     &co, 65536, (abstract_http_coroutine_cb)1,
+                                     &co, 65536,
+                                     (abstract_http_coroutine_cb)(size_t)1,
                                      NULL);
     if (edge_rc == 0)                            /* LCOV_EXCL_LINE */
       abstract_http_coroutine_free(co);          /* LCOV_EXCL_LINE */
